Refactored most caching logic and fixed minor bugs

+++ app/src/main/java/com/myxoz/life/dbwrapper/days/DaysEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "days")
 data class DaysEntity(
     @PrimaryKey(autoGenerate = false) // we’ll provide our own id
-    @ColumnInfo(name = "date") val id: Int,
+    @ColumnInfo(name = "date") val id: Long,
 
     @ColumnInfo(name = "screen_time_ms") val screenTimeMs: Int,
